code cleanup. removed connectivity manager. speakers available in SpeakersFragment
