+++ app/build.gradle
@@ -46,7 +46,6 @@ dependencies {
     compile 'com.android.support.constraint:constraint-layout:1.0.0-alpha3'
     compile "io.reactivex:rxjava:1.1.6"
     compile 'io.reactivex:rxandroid:1.1.0'
-    compile 'com.github.pwittchen:reactivenetwork:0.4.0'
     compile 'com.jakewharton.timber:timber:4.1.2'
     compile 'com.squareup.picasso:picasso:2.5.2'
     compile 'com.jakewharton.threetenabp:threetenabp:1.0.3'
