+++ app/src/hamburg/java/de/devfest/ApplicationModule.java
@@ -6,8 +6,6 @@ import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
-import de.devfest.connectivity.ConnectivityManager;
-import de.devfest.connectivity.DevFestConnectivityManager;
 import de.devfest.data.DevFestManager;
 import de.devfest.data.firebase.FirebaseDevFestManager;
 import de.devfest.user.FirebaseUserManager;
@@ -20,12 +18,6 @@ public class ApplicationModule {
 
     public ApplicationModule(Application application) {
         this.application = application;
-    }
-
-    @Provides
-    @Singleton
-    ConnectivityManager provideConnectivityManager() {
-        return new DevFestConnectivityManager(application);
     }
 
     @Provides
