+++ app/src/main/java/de/devfest/data/DevFestManager.java
@@ -5,8 +5,13 @@ import de.devfest.model.Speaker;
 import de.devfest.model.Track;
 import rx.Observable;
 
-
+/**
+ * TODO: may be split topics here
+ */
 public interface DevFestManager {
+    /**
+     * @return all speakers
+     */
     Observable<Speaker> getSpeakers();
 
     Observable<Speaker> getSpeaker(String uid);
