+++ app/src/main/java/de/devfest/injection/ApplicationComponent.java
@@ -6,12 +6,12 @@ import javax.inject.Singleton;
 
 import dagger.Component;
 import de.devfest.ApplicationModule;
-import de.devfest.connectivity.ConnectivityManager;
 import de.devfest.data.DevFestManager;
 import de.devfest.screens.main.MainActivity;
 import de.devfest.screens.schedule.ScheduleFragment;
 import de.devfest.screens.speakers.SpeakersFragment;
 import de.devfest.screens.user.UserFragment;
+import de.devfest.user.UserManager;
 
 @Singleton
 @Component(modules = ApplicationModule.class)
@@ -25,6 +25,6 @@ public abstract class ApplicationComponent {
     public abstract void inject(ScheduleFragment fragment);
     public abstract void inject(SpeakersFragment fragment);
 
-    abstract ConnectivityManager connectivityManager();
     abstract DevFestManager devFestManager();
+    abstract UserManager userManager();
 }
