+++ app/src/main/java/de/devfest/screens/speakers/SpeakersFragment.java
@@ -2,6 +2,7 @@ package de.devfest.screens.speakers;
 
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,9 +13,10 @@ import javax.inject.Inject;
 import de.devfest.R;
 import de.devfest.databinding.FragmentScheduleBinding;
 import de.devfest.injection.ApplicationComponent;
+import de.devfest.model.Speaker;
 import de.devfest.mvpbase.BaseFragment;
 
-public class SpeakersFragment extends BaseFragment<SpeakersView, SpeakersPresenter> {
+public class SpeakersFragment extends BaseFragment<SpeakersView, SpeakersPresenter> implements SpeakersView {
 
     @Inject
     SpeakersPresenter presenter;
@@ -38,5 +40,16 @@ public class SpeakersFragment extends BaseFragment<SpeakersView, SpeakersPresent
     protected SpeakersPresenter inject(ApplicationComponent component) {
         component.inject(this);
         return presenter;
+    }
+
+
+    @Override
+    public void onSpeakerAvailable(@NonNull Speaker speaker) {
+        // TODO: add to adapter and show
+    }
+
+    @Override
+    public void onError(Throwable error) {
+
     }
 }
