+++ app/src/main/java/de/devfest/screens/user/UserPresenter.java
@@ -4,7 +4,6 @@ import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 
 import javax.inject.Inject;
 
-import de.devfest.connectivity.ConnectivityManager;
 import de.devfest.mvpbase.BasePresenter;
 import de.devfest.user.UserManager;
 import rx.android.schedulers.AndroidSchedulers;
@@ -12,24 +11,16 @@ import rx.schedulers.Schedulers;
 
 public class UserPresenter extends BasePresenter<UserView> {
 
-    private final ConnectivityManager connectivityManager;
     private final UserManager userManager;
 
     @Inject
-    public UserPresenter(ConnectivityManager connectivityManager, UserManager userManager) {
-        this.connectivityManager = connectivityManager;
+    public UserPresenter(UserManager userManager) {
         this.userManager = userManager;
     }
 
     @Override
     public void onResume() {
         super.onResume();
-//        untilOnPause(connectivityManager
-//                .observeInternetConnectivity()
-//                .observeOn(AndroidSchedulers.mainThread())
-//                .subscribe(connected -> {
-//                    getView().showConnectivity(connected);
-//                }));
         userManager.getCurrentUser()
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
